Fix: 2 bytes for ATQA

+++ Firmware/Chameleon-Mini/Application/MifareClassic.c
@@ -910,12 +910,12 @@ void MifareClassicSetUid(ConfigurationUidType Uid)
     }
 }
 
-void MifareClassicGetAtqa(uint8_t * Atqa)
+void MifareClassicGetAtqa(uint16_t * Atqa)
 {
 	*Atqa = CardATQAValue;	
 }
 
-void MifareClassicSetAtqa(uint8_t Atqa)
+void MifareClassicSetAtqa(uint16_t Atqa)
 {
 	CardATQAValue = Atqa;
 }
